refactor: moves function to dedicated util file
Relates #36
